Reorganized entities
