+++ app/src/main/kotlin/fobo66/exchangecourcesbelarus/ui/Navigation.kt
@@ -40,9 +40,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
-import dev.fobo66.valiutchik.presentation.MainScreenState
 import dev.fobo66.valiutchik.presentation.MainViewModel
 import dev.fobo66.valiutchik.presentation.OpenSourceLicensesViewModel
+import dev.fobo66.valiutchik.presentation.entity.MainScreenState
 import fobo66.exchangecourcesbelarus.R
 import fobo66.exchangecourcesbelarus.ui.licenses.OpenSourceLicensesScreen
 import fobo66.exchangecourcesbelarus.ui.main.BestRatesGrid
