+++ app/src/main/kotlin/fobo66/exchangecourcesbelarus/ui/licenses/OpenSourceLicensesScreen.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
-import dev.fobo66.valiutchik.presentation.LicensesState
+import dev.fobo66.valiutchik.presentation.entity.LicensesState
 import fobo66.exchangecourcesbelarus.R
 import fobo66.exchangecourcesbelarus.ui.ProgressIndicator
 import fobo66.exchangecourcesbelarus.ui.main.SecondaryTopBar
