+++ presentation/src/commonMain/kotlin/dev/fobo66/valiutchik/presentation/entity/MainScreenStateTrigger.kt
@@ -14,22 +14,9 @@
  *    limitations under the License.
  */
 
-package dev.fobo66.valiutchik.presentation
+package dev.fobo66.valiutchik.presentation.entity
 
-import androidx.compose.runtime.Immutable
-import androidx.compose.runtime.Stable
-
-@Stable
-sealed class MainScreenState {
-    @Immutable
-    object Initial : MainScreenState()
-
-    @Immutable
-    object Loading : MainScreenState()
-
-    @Immutable
-    object LoadedRates : MainScreenState()
-
-    @Immutable
-    object Error : MainScreenState()
-}
+data class MainScreenStateTrigger(
+    val isRefreshTriggered: Boolean,
+    val isLocationAvailable: Boolean?
+)
